feat(UiManager) : 입출력받는 클래스 따로 분리
